Initial Commit of pulserDCRcalculator.cc
Draws the input range of runs in pulser event data, then implements an increasingly constraining cut on the data until 99% and 96% of the data is remaining. Then, this will draw the pulser distribution in one dimension with vertical lines indicating the cuts at 99% and 96% of the data, with specific information about the acceptance rates and line positions located in the graph's legend. This gives an accurate estimate on what shift in the distribution is going to be a significant factor of our real acceptance rates.